backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
python-multipart
langchain
langchain-openai
langchain-google-genai
